refactor: extract TestTerrainBase class
